Minor fix: integer type in block size round-up.

+++ src/Focal_Setup.f90
@@ -503,7 +503,7 @@ submodule (Focal) Focal_Setup
   module procedure fclLaunchKernel !(kernel,a0,a1,a2,a3,a4,a5,a6,a7,a8,a9,&
                                       ! a10,a11,a12,a13,a14,a15,a16,a17,a18,a19)
 
-    integer :: i, nBlocki
+    integer(c_size_t) :: i, nBlocki
     integer(c_int32_t) :: errcode
     type(fclCommandQ), pointer :: cmdQ
     type(c_ptr) :: localSizePtr
